+++ block_mul/double/bm_4warp_168.cu
@@ -4,16 +4,16 @@
 #include <assert.h>
 
 int N;
-float *A, *B, *C_true;
-float *C;
+double *A, *B, *C_true;
+double *C;
 #define Bs 32
 
-void load_matrix(float**, int, const char *filename);
-void correctness_check(const float*, const float*, int);
-void show(float*, int);
+void load_matrix(double**, int, const char *filename);
+void correctness_check(const double*, const double*, int);
+void show(double*, int);
 
 
-__inline__ __device__ float warpReduceSum(float val) {
+__inline__ __device__ double warpReduceSum(double val) {
     #pragma unroll
     for (int offset = 16; offset > 0; offset >>= 1)
         val += __shfl_down_sync(0xffffffff, val, offset);
@@ -25,13 +25,13 @@ __inline__ __device__ float warpReduceSum(float val) {
 __global__ void block_mul_kernel(
     int N,
     int n,
-    float *d_A,
-    float *d_B,
-    float *d_C
+    double *d_A,
+    double *d_B,
+    double *d_C
 ){
-    extern __shared__ float share[];
-    float *A_block = share;
-    float *B_block = share + Bs * Bs;
+    extern __shared__ double share[];
+    double *A_block = share;
+    double *B_block = share + Bs * Bs;
 
     // block index 
     int bx = blockIdx.x;
@@ -60,8 +60,8 @@ __global__ void block_mul_kernel(
     int col1 = ly + blockDim.x;
 
 
-    float c00=0, c01=0, c10=0, c11=0, c20=0, c21=0, c30=0, c31=0;
-    float a0=0, a1=0, a2=0, a3=0, b0=0, b1=0;
+    double c00=0, c01=0, c10=0, c11=0, c20=0, c21=0, c30=0, c31=0;
+    double a0=0, a1=0, a2=0, a3=0, b0=0, b1=0;
 
    // compute c
     for(int bk = 0; bk< n ; bk++){
@@ -137,15 +137,15 @@ int main(int argc, char* argv[]){
     load_matrix(&A, N, a_filename);
     load_matrix(&B, N, b_filename);
     load_matrix(&C_true, N, c_true_filename); 
-    C = (float*)malloc(N * N * sizeof(float));
+    C = (double*)malloc(N * N * sizeof(double));
 
     //int Bs = 32; // Block size
     int n = N / Bs;
     assert(N % Bs ==0);
 
     // cudaMalloc
-    float *d_A, *d_B, *d_C;
-    size_t gmem = N * N *  sizeof(float);
+    double *d_A, *d_B, *d_C;
+    size_t gmem = N * N *  sizeof(double);
     assert(gmem * 3 < prop.totalGlobalMem);
     cudaMalloc((void **)&d_A, gmem);
     cudaMalloc((void **)&d_B, gmem);
@@ -157,7 +157,7 @@ int main(int argc, char* argv[]){
 
 
     //launch kernel
-    size_t shmem = Bs * Bs * 2 * sizeof(float);
+    size_t shmem = Bs * Bs * 2 * sizeof(double);
     block_mul_kernel<<<dim3(n,n), dim3(16,8), shmem>>>(
         N,
         n,
@@ -181,9 +181,9 @@ int main(int argc, char* argv[]){
     return 0;
 }
 
-void load_matrix(float **mat_ptr, int N_dim, const char *filename) {
+void load_matrix(double **mat_ptr, int N_dim, const char *filename) {
     long long size = (long long)N_dim * N_dim;
-    size_t bytes = size * sizeof(float);
+    size_t bytes = size * sizeof(double);
     
     FILE *file = fopen(filename, "rb"); 
     if (file == NULL) {
@@ -191,14 +191,14 @@ void load_matrix(float **mat_ptr, int N_dim, const char *filename) {
         exit(EXIT_FAILURE);
     }
 
-    *mat_ptr = (float *)malloc(bytes);
+    *mat_ptr = (double *)malloc(bytes);
     if (*mat_ptr == NULL) {
         perror("Host malloc failed in load_matrix");
         fclose(file);
         exit(EXIT_FAILURE);
     }
 
-    size_t read_count = fread(*mat_ptr, sizeof(float), size, file);
+    size_t read_count = fread(*mat_ptr, sizeof(double), size, file);
     if (read_count != size) {
         fprintf(stderr, "Error: Read incomplete from %s. Expected %lld elements, read %zu.\n", 
                 filename, size, read_count);
@@ -209,14 +209,14 @@ void load_matrix(float **mat_ptr, int N_dim, const char *filename) {
     fclose(file);
 }
 
-void correctness_check(const float *C_true, const float *C_result, int N){
+void correctness_check(const double *C_true, const double *C_result, int N){
     int mismatch_count = 0;
-    float tol = 5e-3f;
+    double tol = 5e-3f;
     long long sz = (long long)N * N;
-    float max_error = 0.0f;
+    double max_error = 0.0f;
     
     for(long long i=0; i < sz; i++){
-        float error = fabsf(C_result[i] - C_true[i]);
+        double error = fabsf(C_result[i] - C_true[i]);
         if (error > max_error) max_error = error;
         
         if (error > tol){
@@ -237,7 +237,7 @@ void correctness_check(const float *C_true, const float *C_result, int N){
     }
 }
 
-void show(float* M, int N){
+void show(double* M, int N){
     for(int i = 0; i < N; i++){
         for(int j = 0; j < N ;j++){
             printf("%.0f, ", M[i * N+ j]);
